+++ choco-solver/src/main/java/org/chocosolver/solver/explanations/Rules.java
@@ -31,23 +31,52 @@ package org.chocosolver.solver.explanations;
 
 import gnu.trove.set.TIntSet;
 import gnu.trove.set.hash.TIntHashSet;
+import org.chocosolver.solver.constraints.Propagator;
+import org.chocosolver.solver.variables.Variable;
 
 import java.util.BitSet;
 
 /**
- * A set of rules for the rulestore
+ * A set of rules for {@link RuleStore}
  * <p>
  * Created by cprudhom on 17/03/15.
  * Project: choco.
+ * @author Charles Prud'homme
  */
 public class Rules {
 
+    /**
+     * no entry value for vmRules
+     */
     private static final int NO_ENTRY = Integer.MIN_VALUE;
-    private BitSet paRules; // rules for propagator activation
-    private BitSet vmRules; // rules for propagator activation
-    private int[] vmMasks;    // rules for variable modification
-    private NoIteratorIntHashSet[] remVal;    // store value removal when necessary
 
+    /**
+     * Stores index of propagators ({@link Propagator#getId()}) involved into an explanation
+     */
+    private BitSet paRules;
+
+    /**
+     * Stores index of variables ({@link Variable#getId()}) involved into an explanation
+     */
+    private BitSet vmRules;
+
+    /**
+     * Stores for variables in {@link #vmRules} the type of modifications related to an explanation.
+     * Possible masks are: REM or LB or UB or MOD.
+     */
+    private int[] vmMasks;
+
+    /**
+     * Stores value removals of a given variable ({@link Variable#getId()}), when needed
+     */
+    private NoIteratorIntHashSet[] remVal;
+
+    /**
+     * Creates a set of rules for {@link RuleStore}.
+     * It stores events related to an explanation.
+     * @param i1 initial size of {@link #vmMasks}
+     * @param i2 initial size of {@link #remVal}
+     */
     public Rules(int i1, int i2) {
         this.paRules = new BitSet();
         this.vmRules = new BitSet();
@@ -96,10 +125,14 @@ public class Rules {
         return NO_ENTRY;
     }
 
+    /**
+     * Makes sure that {@link #remVal} is large enough to store data
+     * @param size excepted size remVal
+     */
     private void ensureRemvalCapacity(int size) {
         if (size >= remVal.length) {
             int nsize = Math.max(size, remVal.length * 3 / 2 + 1);
-            TIntSet[] tmp = remVal;
+            NoIteratorIntHashSet[] tmp = remVal;
             remVal = new NoIteratorIntHashSet[nsize];
             System.arraycopy(tmp, 0, remVal, 0, tmp.length);
         }
@@ -121,6 +154,10 @@ public class Rules {
         return remVal[vid];
     }
 
+    /**
+     * Makes sure that {@link #vmMasks} is large enough to store data
+     * @param size excepted size of vmMasks
+     */
     private void ensureRulesCapacity(int size) {
         if (size >= vmMasks.length) {
             int nsize = Math.max(size, vmMasks.length * 3 / 2 + 1);
@@ -168,6 +205,7 @@ public class Rules {
      * @param i1  lower bound of the interval (included)
      * @param i2  upper bound of the interval (included)
      * @param vid variable id
+     * @return <tt>true</tt> if intersection is not empty
      */
     public boolean intersect(int i1, int i2, int vid) {
         assert vid < remVal.length && remVal[vid].size() > 0;
@@ -226,6 +264,9 @@ public class Rules {
 
         public NoIteratorIntHashSet(int initial_capacity, float load_factor, int no_entry_value) {
             super(initial_capacity, load_factor, no_entry_value);
+        }
+
+        public NoIteratorIntHashSet() {
         }
 
         /**