+++ CHANGES.md
@@ -14,6 +14,11 @@ NEXT MILESTONE
 - update javadoc (in particular: #347)
 - add default name to Solver + setter, modify measures printing to include the name.
 
+
+#### Explanations
+
+- refactor `PropNogoods` to deal with generalized no-goods
+
 #### Bug fixes: 
 
 \#346, #348, #349, #350